+++ system/src/chat/hooks.mjs
@@ -11,21 +11,13 @@
  * @returns {object[]}          The extended options Array including new context choices
  */
 export function addChatMessageContextOptions(html, options) {
+
 	const canApplyDamage = li => {
 		const message = getMessage(li);
 
 		return game.user.isGM
 			&& canvas.tokens?.controlled.length
-			&& (_chatMessageIsBasicRoll(message)
-				|| _chatMessageIsDamageCard(message));
-	};
-
-	const canApplySecondaryDamage = li => {
-		const message = getMessage(li);
-
-		return game.user.isGM
-			&& canvas.tokens?.controlled.length
-			&& (_chatMessageIsDamageCardSecondary(message));
+			&& _chatMessageIsDamageCard(message);
 	};
 
 	options.push(
@@ -40,18 +32,6 @@ export function addChatMessageContextOptions(html, options) {
 			icon: '<i class="fas fa-user-plus"></i>',
 			condition: canApplyDamage,
 			callback: li => applyChatCardDamage(li, -1),
-		},
-		{
-			name: game.i18n.localize("SHADOWDARK.chat_card.context.apply_damage_secondary"),
-			icon: '<i class="fas fa-user-minus"></i>',
-			condition: canApplySecondaryDamage,
-			callback: li => applyChatCardDamageSecondary(li, 1),
-		},
-		{
-			name: game.i18n.localize("SHADOWDARK.chat_card.context.apply_healing_secondary"),
-			icon: '<i class="fas fa-user-plus"></i>',
-			condition: canApplySecondaryDamage,
-			callback: li => applyChatCardDamageSecondary(li, -1),
 		}
 	);
 
@@ -77,63 +57,15 @@ function getMessage(element) {
  * @returns {Promise}
  */
 function applyChatCardDamage(li, multiplier) {
-
 	const message = getMessage(li);
-	let roll;
+	const roll = message.rolls[1];
 
-	// There are two version of this check.
-	// Version #1, this is a single roll
-
-	if (_chatMessageIsBasicRoll(message)) {
-		roll = message.rolls[0];
-	}
-	else if (_chatMessageIsDamageCard(message)) {
-		roll = message?.flags?.shadowdark?.rolls?.damage?.roll;
-	}
-	else {
-		return;
-	}
+	if (!roll) return;
 
 	return Promise.all(canvas.tokens.controlled.map(t => {
 		const a = t.actor;
 		return a.applyDamage(roll.total, multiplier);
 	}));
-}
-
-/**
- * Apply secondary rolled dice damage to the token or tokens which are currently controlled.
- * The multipliers allows for damage to be scaled for healing, or other modifications.
- * Specifically used for damage cards with two outputs, such at versatile.
- *
- * @param {HTMLElement} li      The chat entry which contains the roll data
- * @param {number} multiplier   A damage multiplier to apply to the rolled damage.
- * @returns {Promise}
- */
-function applyChatCardDamageSecondary(li, multiplier) {
-
-	const message = getMessage(li);
-
-	if (!_chatMessageIsDamageCardSecondary(message)) {
-		return;
-	}
-
-	let roll = message?.flags?.shadowdark?.rolls?.secondaryDamage?.roll;
-
-	return Promise.all(canvas.tokens.controlled.map(t => {
-		const a = t.actor;
-		return a.applyDamage(roll.total, multiplier);
-	}));
-}
-
-
-/**
- * Identifies basic ChatMessage rolls like `/r d6`
- * @param {ChatMessage} message
- * @returns
- */
-function _chatMessageIsBasicRoll(message) {
-	return message?.isRoll
-		&& message?.rolls[0];
 }
 
 /**
@@ -142,17 +74,5 @@ function _chatMessageIsBasicRoll(message) {
  * @returns
  */
 function _chatMessageIsDamageCard(message) {
-	return message?.flags?.shadowdark?.isRoll
-		&& (message?.flags?.shadowdark?.rolls?.damage);
-}
-
-/**
- * Identifies our custom Attack + Damage card with secondary damage roll
- * Clasically, this is a versatile weapon
- * @param {ChatMessage} message
- * @returns
- */
-function _chatMessageIsDamageCardSecondary(message) {
-	return message?.flags?.shadowdark?.isRoll
-		&& message?.flags?.shadowdark?.rolls?.secondaryDamage;
+	return message?.flags?.shadowdark?.rollConfig?.damageRoll ? true : false;
 }