+++ system/src/dice/dice.mjs
@@ -20,12 +20,31 @@ export function applyAdvantage(formula, adv) {
 }
 
 /**
+ * Applies critical hit to the provided roll formula
+ * @param {string} formula // roll formula
+ * @returns {string} // new roll formula
+ */
+export function applyCriticalHit(formula, multiplier) {
+	return formula.replace(/(\d*)d(\d+[a-z0-9]*)/ig, function(match) {
+		if (match.startsWith("d")) {
+			let count = Number(multiplier);
+			return `${count}${match}`; // match starting with a "d" means a single die
+		}
+		else {
+			let diceCountMatch = match.match(/(\d)(d[a-z0-9]+)/);
+			let count = Number(diceCountMatch[1]) * Number(multiplier);
+			return `${count}${diceCountMatch[2]}`;
+		}
+	});
+}
+
+/**
  * Applies exploding dice to the provided roll formula
  * @param {string} formula // roll formula
  * @returns {string} // new roll formula
  */
 export function applyExploding(formula) {
-	return formula.replace(/^(\d*)d(\d+)/, function(match, dice, sides) {
+	return formula.replace(/(\d*)d(\d+[a-z0-9]*)/i, function(match) {
 		return `${match}x`;
 	});
 }
@@ -69,8 +88,8 @@ export function createToolTip(name, value, prefix="+", key="") {
 export function formatBonus(bonus) {
 	if (typeof bonus === "number") {
 		if (bonus === 0) return "";
-		if (bonus > 0) return ` +${bonus}`;
-		if (bonus < 0) return ` ${bonus}`;
+		if (bonus > 0) return ` + ${bonus}`;
+		if (bonus < 0) return ` - ${Math.abs(bonus)}`;
 	}
 	return bonus;
 }
@@ -128,6 +147,10 @@ export async function roll(config, rolldata={}) {
 	}
 
 	if (config.type === "damage") {
+		// double base damage on critical hit
+		if (config.criticalHit) {
+			config.formula = applyCriticalHit(config.formula, config.criticalMultiplier);
+		}
 		// apply momentum mode
 		if (game.settings.get("shadowdark", "useMomentumMode")) {
 			config.formula = applyExploding(config.formula);
@@ -147,23 +170,15 @@ export async function rollDamageFromMessage(msg) {
 	if (!config.damageRoll?.formula || msg.getRoll("damage")) return false;
 	const actor = game.actors.get(config.actorId);
 	if (!actor) return; // TODO Error message
-	config.damageRoll.type = "damage";
-	const damageRoll = await roll(config.damageRoll, actor.getRollData());
-	config.damageRoll.html = await damageRoll.render();
 
-	// Generate template data a new content
-	const template = "systems/shadowdark/templates/chat/roll-card.hbs";
-	const templateData = {...config};
-	templateData.actor = actor;
-	templateData.mainRoll.html = await msg.getRoll("main").render();
-	templateData.damageRoll.html = await damageRoll.render();
-	if (config.itemUuid) {
-		templateData.item = await fromUuid(config.itemUuid);
-	}
-	if (config.targetUuid) {
-		templateData.target = await fromUuid(config.targetUuid);
-	}
-	const content = await foundry.applications.handlebars.renderTemplate(template, templateData);
+	const mainRoll = msg.getRoll("main");
+
+	config.damageRoll.type = "damage";
+	config.damageRoll.criticalHit = mainRoll.criticalSuccess;
+	config.damageRoll.criticalMultiplier = mainRoll.options.criticalMultiplier;
+	const damageRoll = await roll(config.damageRoll, actor.getRollData());
+
+	const content = await shadowdark.chat.renderRollCard(config, [mainRoll, damageRoll]);
 
 	// update message with new roll and content
 	await msg.update({rolls: [...msg.rolls, damageRoll]});
@@ -234,6 +249,9 @@ export async function rollFromConfig(config) {
 
 	if (mainRoll.success && config?.damageRoll?.formula) {
 		// await rollDamageFromMessage(message);
+		config.damageRoll.needed = true;
+	}
+	else if (!config.targetUuid && config?.damageRoll?.formula) {
 		config.damageRoll.needed = true;
 	}