+++ system/src/models/_ActorBaseSD.mjs
@@ -133,16 +133,20 @@ export class ActorBaseSD extends foundry.abstract.TypeDataModel {
 
 	_getAbilityModifier(ability) {
 		if (!CONFIG.SHADOWDARK.ABILITY_KEYS.includes(ability)) return;
+
 		const modifier = this.abilities[ability].mod;
-		const tooltip = [];
-		if (modifier !==0) {
-			tooltip.push(shadowdark.dice.createToolTip(
+
+		let tooltip = "";
+		if (modifier !== 0) {
+			tooltip = shadowdark.dice.createToolTip(
 				game.i18n.format(
 					"SHADOWDARK.roll.tooltip.stat_bonus",
-					{stat: CONFIG.SHADOWDARK.ABILITIES_LONG[ability]}),
+					{stat: CONFIG.SHADOWDARK.ABILITIES_LONG[ability]}
+				),
 				modifier
-			));
+			);
 		}
+
 		return {modifier, tooltip};
 	}
 
@@ -164,6 +168,12 @@ export class ActorBaseSD extends foundry.abstract.TypeDataModel {
 		if (baseKey.startsWith("system.roll.melee") || baseKey.startsWith("system.roll.ranged")) {
 			baseKeys.push(
 				baseKey.replace(/^system\.roll\.(melee|ranged)\./, "system.roll.attack.")
+			);
+		}
+		else if (baseKey.startsWith("system.roll.stat.")) {
+			// Also search for the "all" key for stat rolls
+			baseKeys.push(
+				baseKey.replace(/(str|dex|con|int|wis|cha)$/, "all")
 			);
 		}
 
@@ -204,7 +214,9 @@ export class ActorBaseSD extends foundry.abstract.TypeDataModel {
 			if (keys.includes(c.key) || isItem) {
 				// include only selected situational active effects
 				if (e.isSituational) {
-					config.situational.push(e.uuid);
+					if (!config.situational.includes(e.uuid)) {
+						config.situational.push(e.uuid);
+					}
 					if (!config.selected?.includes(e.uuid)) return;
 				}
 
@@ -271,12 +283,13 @@ export class ActorBaseSD extends foundry.abstract.TypeDataModel {
 
 		// generate check formula from ability mod and AE roll bonuses
 		const abilityMod = this._getAbilityModifier(ability);
-		const rollKey = this._getActiveEffectKeys(`roll.${ability}.bonus`, abilityMod.modifier, null, config);
+		const rollKey = this._getActiveEffectKeys(`roll.stat.bonus.${ability}`, abilityMod.modifier, null, config);
 		config.mainRoll.bonus = shadowdark.dice.formatBonus(rollKey.value);
 		config.mainRoll.formula = `${config.mainRoll.base}${config.mainRoll.bonus}`;
+		config.mainRoll.formulaBackup = config.mainRoll.formula;
 
 		// calculate roll advantage
-		const advRollKeyAdv = this._getActiveEffectKeys(`roll.${ability}.advantage`, 0, null, config);
+		const advRollKeyAdv = this._getActiveEffectKeys(`roll.stat.advantage.${ability}`, 0, null, config);
 		config.mainRoll.advantage = advRollKeyAdv.value;
 
 		// generate tooltips