+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - [#1038] Added missing translatable strings
 - [#1152] Add "Momentum Mode" Option *(Ashley Towner)*
 - [#1202] Ammo type should be overridable as part of the attack roll dialog
+- [#1216] Allow Versatile weapons to be switched between 1H and 2H in the inventory so correct damage is displayed/rolled
 
 ## Bugfixes
 - [#1068] Blowgun weapon attack can now be rolled
@@ -17,11 +18,28 @@
 - [#1143] CSS in v12 causing problems, cannot override
 - [#1146] Remaining torch time shown on chat card when using hide timer *(Brittonlj)*
 - [#1194] Compendium migrations skipped when in a Module
+- [#1204] Too much damage taken when applied from chat card
+- [#1205] Damage should always be rolled if no actors have been targetted
+- [#1214] Unable to open Player character sheets directly from compendiums
+- [#1215] No way of rolling 2H damage for versatile weapons currently
+- [#1247] Critical Success attacks do not roll double the dice for damage
 
 ## Chores
 - [#369] Migrated to using data models for actors and items
 - [#1153] Add tip for improving editor auto-complete *(puradox)*
+- [#1167] Migrate Pre-existing Active Effects to new model
 - [#1193] Remove redundant span elements from all compendium text fields.
+- [#1213] Sanity check Quickstart adventure and Pregen characters to ensure the new Active Effects are being used effectively
+- [#1218] Accessing the global "Items" which is now namespaced under foundry.documents.collections.Items
+- [#1219] Accessing the global "Actors" which is now namespaced under foundry.documents.collections.Actors
+- [#1220] Accessing the global "ActorSheet" which is now namespaced under foundry.appv1.sheets.ActorSheet
+- [#1221] Accessing the global "ItemSheet" which is now namespaced under foundry.appv1.sheets.ItemSheet
+- [#1222] Accessing the global "loadTemplates" which is now namespaced under foundry.applications.handlebars.loadTemplates
+- [#1223] Accessing the global "TextEditor" which is now namespaced under foundry.applications.ux.TextEditor.implementation
+- [#1225] TableResult#documentId is deprecated. Consult TableResult#uuid instead
+- [#1229] Update Active Effects drop handlers that use the REPLACEME placeholder to trigger configuration to use new AE schema
+- [#1230] Merge German translation updates from Crowdin
+- [#1237] Merge Spanish translation updates from Crowdin
 
 *Many thanks to **Ashley Towner <git@ashtowner.com>**, **andrewbeard**, **Brittonlj**, **johncarney**, and **puradox** for contributing to this release*